benchmark/requirements.txt
mzn-bench
mzn-bench[plotting]
python-dotenv